Maven Checkstyle fixes and open-api-tool configuration changes to pick the module.
