+++ component/org.wso2.carbon.identity.api.otp.service.smsotp/src/gen/java/org/wso2/carbon/identity/api/otp/service/smsotp/dto/OTPValidationRequest.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.identity.api.otp.service.smsotp.dto;
+package org.wso2.carbon.identity.api.otp.service.smsotp.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModelProperty;
@@ -25,24 +25,26 @@ import java.util.Objects;
 import javax.validation.Valid;
 
 /**
- * OTPGenerationRequest DTO to get the request OTP generation.
+ * OTPValidationRequest DTO to validate OTP.
  */
-public class OTPGenerationRequest {
+public class OTPValidationRequest {
 
     private String userId;
+    private String transactionId;
+    private String smsOTP;
 
     /**
-     * SCIM Id of the user.
+     * SCIM ID of the user.
      *
      * @param userId
      **/
-    public OTPGenerationRequest userId(String userId) {
+    public OTPValidationRequest userId(String userId) {
 
         this.userId = userId;
         return this;
     }
 
-    @ApiModelProperty(example = "8b1cc9c4-b671-448a-a334-5afe838a4a3b", value = "SCIM Id of the user")
+    @ApiModelProperty(example = "8b1cc9c4-b671-448a-a334-5afe838a4a3b", value = "SCIM ID of the user")
     @JsonProperty("userId")
     @Valid
     public String getUserId() {
@@ -55,6 +57,54 @@ public class OTPGenerationRequest {
         this.userId = userId;
     }
 
+    /**
+     * Unique identifier of the generated OTP.
+     *
+     * @param transactionId
+     **/
+    public OTPValidationRequest transactionId(String transactionId) {
+
+        this.transactionId = transactionId;
+        return this;
+    }
+
+    @ApiModelProperty(value = "Unique identifier of the generated OTP")
+    @JsonProperty("transactionId")
+    @Valid
+    public String getTransactionId() {
+
+        return transactionId;
+    }
+
+    public void setTransactionId(String transactionId) {
+
+        this.transactionId = transactionId;
+    }
+
+    /**
+     * SMS OTP.
+     *
+     * @param smsOTP
+     **/
+    public OTPValidationRequest smsOTP(String smsOTP) {
+
+        this.smsOTP = smsOTP;
+        return this;
+    }
+
+    @ApiModelProperty(value = "SMS OTP")
+    @JsonProperty("smsOTP")
+    @Valid
+    public String getSmsOTP() {
+
+        return smsOTP;
+    }
+
+    public void setSmsOTP(String smsOTP) {
+
+        this.smsOTP = smsOTP;
+    }
+
     @Override
     public boolean equals(java.lang.Object o) {
 
@@ -64,23 +114,27 @@ public class OTPGenerationRequest {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        OTPGenerationRequest otPGenerationRequest = (OTPGenerationRequest) o;
-        return Objects.equals(this.userId, otPGenerationRequest.userId);
+        OTPValidationRequest otPValidationRequest = (OTPValidationRequest) o;
+        return Objects.equals(this.userId, otPValidationRequest.userId) &&
+                Objects.equals(this.transactionId, otPValidationRequest.transactionId) &&
+                Objects.equals(this.smsOTP, otPValidationRequest.smsOTP);
     }
 
     @Override
     public int hashCode() {
 
-        return Objects.hash(userId);
+        return Objects.hash(userId, transactionId, smsOTP);
     }
 
     @Override
     public String toString() {
 
         StringBuilder sb = new StringBuilder();
-        sb.append("class OTPGenerationRequest {\n");
+        sb.append("class OTPValidationRequest {\n");
 
         sb.append("    userId: ").append(toIndentedString(userId)).append("\n");
+        sb.append("    transactionId: ").append(toIndentedString(transactionId)).append("\n");
+        sb.append("    smsOTP: ").append(toIndentedString(smsOTP)).append("\n");
         sb.append("}");
         return sb.toString();
     }
