+++ component/org.wso2.carbon.identity.api.otp.service.smsotp/src/gen/java/org/wso2/carbon/identity/api/otp/service/smsotp/dto/OTPValidationResponse.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.identity.api.otp.service.smsotp.dto;
+package org.wso2.carbon.identity.api.otp.service.smsotp.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModelProperty;
@@ -25,59 +25,58 @@ import java.util.Objects;
 import javax.validation.Valid;
 
 /**
- * OTPGenerationResponse DTO to get the response of OTP generation.
+ * OTPValidationResponse DTO to get the validate OTP response.
  */
-public class OTPGenerateResponse {
+public class OTPValidationResponse {
 
-    private String transactionId;
-    private String smsOTP;
+    private Boolean isValid;
+    private String userId;
 
     /**
-     * Unique identifier of the generated OTP.
+     * SMS OTP validated successfully.
      *
-     * @param transactionId     TransactionId of the OTP.
+     * @param  isValid
      **/
-    public OTPGenerateResponse transactionId(String transactionId) {
+    public OTPValidationResponse isValid(Boolean isValid) {
 
-        this.transactionId = transactionId;
+        this.isValid = isValid;
         return this;
     }
 
-    @ApiModelProperty(value = "Unique identifier of the generated OTP")
-    @JsonProperty("transactionId")
+    @ApiModelProperty(value = "SMS OTP validated successfully")
+    @JsonProperty("isValid")
     @Valid
-    public String getTransactionId() {
+    public Boolean getIsValid() {
 
-        return transactionId;
+        return isValid;
     }
 
-    public void setTransactionId(String transactionId) {
+    public void setIsValid(Boolean isValid) {
 
-        this.transactionId = transactionId;
+        this.isValid = isValid;
     }
 
     /**
-     * The generated OTP.
-     *
-     * @param smsOTP    generated OTP.
+     * SCIM ID of the user which the token issued and validated.
      **/
-    public OTPGenerateResponse smsOTP(String smsOTP) {
+    public OTPValidationResponse userId(String userId) {
 
-        this.smsOTP = smsOTP;
+        this.userId = userId;
         return this;
     }
 
-    @ApiModelProperty(value = "The generated OTP")
-    @JsonProperty("smsOTP")
+    @ApiModelProperty(example = "8b1cc9c4-b671-448a-a334-5afe838a4a3b", value = "SCIM ID of the user which the token " +
+            "issued and validated")
+    @JsonProperty("userId")
     @Valid
-    public String getSmsOTP() {
+    public String getUserId() {
 
-        return smsOTP;
+        return userId;
     }
 
-    public void setSmsOTP(String smsOTP) {
+    public void setUserId(String userId) {
 
-        this.smsOTP = smsOTP;
+        this.userId = userId;
     }
 
     @Override
@@ -89,25 +88,25 @@ public class OTPGenerateResponse {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        OTPGenerateResponse otPGenerateResponse = (OTPGenerateResponse) o;
-        return Objects.equals(this.transactionId, otPGenerateResponse.transactionId) &&
-                Objects.equals(this.smsOTP, otPGenerateResponse.smsOTP);
+        OTPValidationResponse otPValidationResponse = (OTPValidationResponse) o;
+        return Objects.equals(this.isValid, otPValidationResponse.isValid) &&
+                Objects.equals(this.userId, otPValidationResponse.userId);
     }
 
     @Override
     public int hashCode() {
 
-        return Objects.hash(transactionId, smsOTP);
+        return Objects.hash(isValid, userId);
     }
 
     @Override
     public String toString() {
 
         StringBuilder sb = new StringBuilder();
-        sb.append("class OTPGenerateResponse {\n");
+        sb.append("class OTPValidationResponse {\n");
 
-        sb.append("    transactionId: ").append(toIndentedString(transactionId)).append("\n");
-        sb.append("    smsOTP: ").append(toIndentedString(smsOTP)).append("\n");
+        sb.append("    isValid: ").append(toIndentedString(isValid)).append("\n");
+        sb.append("    userId: ").append(toIndentedString(userId)).append("\n");
         sb.append("}");
         return sb.toString();
     }
